1. Se finalizó la edición de una categoría. 2. Se reutilizó el formulario de la creación, para la edición de categorías. 3. Se documentó lo trabajado, servicios, archivos .ts

+++ src/app/admin/components/brands-list/brands-list.component.ts
@@ -27,10 +27,11 @@ export class BrandsListComponent implements OnInit {
   }
 
   deleteBrand(id: number): void {
-    this.brandService.deleteBrand(id)
-    .subscribe(() => {
-      this.fetchAllBrands();
-    });
+    if (confirm('¿Estás seguro?')) {
+      this.brandService.deleteBrand(id)
+      .subscribe(() => {
+        this.fetchAllBrands();
+      });
+    }
   }
-
 }
